Converted timestamp to int and added altitude to location

+++ MobileClients/Android/HUB/src/ch/ethz/coss/nervousnet/sensors/model/LocationReading.java
@@ -37,7 +37,7 @@ public class LocationReading extends SensorReading {
 
 	public void readFromParcel(Parcel in) {
 
-		timestamp = in.readLong();
+		timestamp = in.readInt();
 		latnLong = in.createDoubleArray();
 	}
